+++ backend/loan-management-api/loan-backend/src/main/java/com/patrick/fintech/loan_backend/model/Borrower.java
@@ -9,8 +9,6 @@ import jakarta.validation.constraints.Pattern;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
-
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -108,9 +106,9 @@ public class Borrower {
     private String employerName;
     private String employmentType;       // PERMANENT, CONTRACT, SELF_EMPLOYED, UNEMPLOYED
     private String jobTitle;
-    private BigDecimal monthlyIncome;
-    private BigDecimal monthlyExpenses;
-    private BigDecimal netWorth;
+    private Double monthlyIncome;
+    private Double monthlyExpenses;
+    private Double netWorth;
 
     // Credit
     private Integer creditScore;
